super Bot

+++ src/main/java/programbot/dao/Dao.java
@@ -5,10 +5,12 @@ import org.hibernate.HibernateException;
 import org.hibernate.ObjectNotFoundException;
 
 public interface Dao<Type> {
-    void save(Type entity)throws HibernateException;
-    Type findById(int id)throws ObjectNotFoundException;
-    void update(Type entity)throws HibernateException;
-    void delete(Type entity)throws HibernateException;
-  //  void rollback()throws HibernateException;
+    void save(Type entity) throws HibernateException;
+
+    Type findById(Long id) throws ObjectNotFoundException;
+
+    void update(Type entity) throws HibernateException;
+
+    void delete(Type entity) throws HibernateException;
 
 }
